requirements.txt
Flask==2.0.1
flask-cors==3.0.10
openai==0.28
python-dotenv==0.19.2
werkzeug==2.0.3
whoosh==2.7.4
beautifulsoup4==4.11.1
gunicorn==20.1.0